backend/core/tools/web_search.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

"""
ARIES - Web搜索工具
实现网络搜索功能，获取外部信息
"""

import os
import json
import logging
import requests
from typing import Dict, List, Any, Optional

class WebSearch:
    """Web搜索类，用于获取外部信息"""
    
    def __init__(self, api_key: str = None):
        """初始化Web搜索工具
        
        Args:
            api_key: 搜索API密钥
        """
        self.api_key = api_key
        self.logger = logging.getLogger("aries_web_search")
    
    def search(self, query: str, num_results: int = 5) -> List[Dict[str, Any]]:
        """执行网络搜索
        
        Args:
            query: 搜索查询
            num_results: 返回结果数量
            
        Returns:
            搜索结果列表
        """
        try:
            self.logger.info(f"执行Web搜索: {query}")
            
            # 检查API密钥
            if not self.api_key:
                self.logger.warning("未提供搜索API密钥，使用模拟搜索")
                return self._mock_search(query, num_results)
            
            # 使用搜索API
            # 这里应该实现实际的搜索API调用
            # 以下是使用Google Custom Search API的示例
            url = "https://www.googleapis.com/customsearch/v1"
            params = {
                "key": self.api_key,
                "cx": "YOUR_SEARCH_ENGINE_ID",  # 需要替换为实际的搜索引擎ID
                "q": query,
                "num": min(num_results, 10)  # Google API最多返回10个结果
            }
            
            response = requests.get(url, params=params)
            
            if response.status_code == 200:
                data = response.json()
                results = []
                
                if "items" in data:
                    for item in data["items"]:
                        result = {
                            "title": item.get("title", ""),
                            "link": item.get("link", ""),
                            "snippet": item.get("snippet", ""),
                            "source": "google"
                        }
                        results.append(result)
                
                return results
            else:
                self.logger.error(f"搜索API请求失败: {response.status_code}")
                return self._mock_search(query, num_results)
                
        except Exception as e:
            self.logger.error(f"执行Web搜索失败: {str(e)}")
            return self._mock_search(query, num_results)
    
    def _mock_search(self, query: str, num_results: int = 5) -> List[Dict[str, Any]]:
        """模拟搜索结果（当API不可用时）
        
        Args:
            query: 搜索查询
            num_results: 返回结果数量
            
        Returns:
            模拟搜索结果列表
        """
        self.logger.info(f"使用模拟搜索: {query}")
        
        # 根据查询关键词生成模拟结果
        mock_results = [
            {
                "title": f"关于 {query} 的结果 1",
                "link": "https://example.com/result1",
                "snippet": f"这是关于 {query} 的模拟搜索结果1。请注意，这是模拟数据，不是实际的搜索结果。",
                "source": "mock"
            },
            {
                "title": f"关于 {query} 的结果 2",
                "link": "https://example.com/result2",
                "snippet": f"这是关于 {query} 的模拟搜索结果2。请配置实际的搜索API以获取真实结果。",
                "source": "mock"
            },
            {
                "title": f"关于 {query} 的结果 3",
                "link": "https://example.com/result3",
                "snippet": f"这是关于 {query} 的模拟搜索结果3。模拟数据仅用于演示目的。",
                "source": "mock"
            },
            {
                "title": f"关于 {query} 的结果 4",
                "link": "https://example.com/result4",
                "snippet": f"这是关于 {query} 的模拟搜索结果4。请在配置文件中设置搜索API密钥。",
                "source": "mock"
            },
            {
                "title": f"关于 {query} 的结果 5",
                "link": "https://example.com/result5",
                "snippet": f"这是关于 {query} 的模拟搜索结果5。实际部署时请替换为真实搜索API。",
                "source": "mock"
            }
        ]
        
        return mock_results[:num_results]
    
    def get_content(self, url: str) -> Dict[str, Any]:
        """获取网页内容
        
        Args:
            url: 网页URL
            
        Returns:
            网页内容
        """
        try:
            self.logger.info(f"获取网页内容: {url}")
            
            # 发送HTTP请求
            headers = {
                "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
            }
            
            response = requests.get(url, headers=headers, timeout=10)
            
            if response.status_code == 200:
                # 使用BeautifulSoup解析HTML
                from bs4 import BeautifulSoup
                soup = BeautifulSoup(response.text, "html.parser")
                
                # 提取标题
                title = soup.title.string if soup.title else ""
                
                # 提取正文内容
                # 移除脚本和样式元素
                for script in soup(["script", "style"]):
                    script.extract()
                
                # 获取文本
                text = soup.get_text(separator="\n", strip=True)
                
                # 清理文本（移除多余空行等）
                lines = [line.strip() for line in text.splitlines() if line.strip()]
                text = "\n".join(lines)
                
                return {
                    "success": True,
                    "url": url,
                    "title": title,
                    "content": text[:5000]  # 限制内容长度
                }
            else:
                return {
                    "success": False,
                    "url": url,
                    "error": f"HTTP错误: {response.status_code}"
                }
                
        except Exception as e:
            self.logger.error(f"获取网页内容失败: {str(e)}")
            return {
                "success": False,
                "url": url,
                "error": str(e)
            }